Count East Asian Wide and Full Fidth chars as two columns.
* regex.c (create_wide_cs): New static function.
(wide_display_char_p): New function.

* regex.h (wide_display_char_p): Declared.

* stream.c (put_string, put_char): Use wide_display_char_p
to determine whether an extra column need be counted.  Also bugfix:
iswprint evidently cannot be relied to work over the entire Unicode
range, at least not in the C locale. Glibc's version and is reporting
valid Japanese characters as unprintable on Ubuntu. As a hack we
instead check for control characters and invert the result: control
chars are unprintable.

* tests/009/json.expected: Updated.