Floating-point constant tightening.
* parser.l (grammar): Change order of rule which recognizes FLODOT with
a one-character trailing context other than a dot, and the rule which
diagnoses trailing junk. The issue is that this order gives the wrong
interpretation to 123.E, treating it as 123.  followed by E rather than
trailing junk, like in the case of 123.0E or 123.B.

* txr.1: Adding the valid example 1.E5.  Removing references to dot as
consing dot.  Fixed documentation which says that 1.E is 1 followed by
a consing dot and E. The wrong behavior in fact produced 1.0 followed
by E. No consing dot semantics.