Reset flex scanner if exception thrown during read.
We must tear down and rebuild the Flex scanner, otherwise
it is left in a bad state causing it to abort.

* parser.c (parser_cleanup): After destroying scanner,
null out the pointer.
(parser_reset): Destroy the Flex scanner, and
instantiate a new one in its place.
(lisp_parse): Set up an unwind catch which will
reset the parser if it was aborted by an exception.

* parser.h (parser_reset): Declared.