Small optimization in finalization.
* gc.c (prepare_finals): In the second pass, only mark
objects that were identified in the first pass as unreachable.
Reachable ones don't require a redundant call to mark_obj.
Also, objects identified as unreachable can be moved to
gen 0. This is something like what the mark_makefresh hack was
trying to do. It only does anything in a full GC pass, because
in an incremental pass, anything identified as unreachable is
necessarily gen 0. It can help prevent some objects that
are revived by finalization from sliding back to gen 1
and hanging around.