Bugfix: make self-path visible to Lisp.
The self-path variable is not bound in TXR Lisp scripts
or the command line.

* txr.c (txr_main): Set self-path to "cmdline-expr" before
command line Lisp evaluations. (For TXR pattern language
syntax invoked in a -c, it continues to be "cmdline").
Set self-path to the file name when reading a Lisp file.
Also, the legacy, undocumented *self-path* is only set
for TXR scripts, and is reverted to being a special variable,
not a global lexical.

* txr.1: Documented self-path in more detail.