Fix misspelled :toinger filter.
The manual documents a filter called :tointeger, but it has
never existed. The keyword was misspelled in the source code
from the beginning as :toinger.

I'm fixing the typo and renaming the filter to :toint
at the same time.

* filter.c (tointeger_k, toint_k): Former renamed
to latter.
(filter_init): Initialize toint_k with correctly
spelled symbol. Register filter under toint_k.

* filter.h (tointeger_k, toint_k): Former renamed
to latter.

* txr.1: Changed mention of :tointeger to :toint.