False positive valgrind error: uninitialized sigset_t.
The issue is that sigset_t is 1024 bits wide on Linux, but
there aren't actually that many signals. Valgrind knows this
and so when sigprocmask returns the old signal set, Valgrind
only marks a portion of it as initialized, and not the
entire 1024 bits. When this sigset_t is later passed into
sig_set again, we do a memcmp on all 1024 bits and Valgrind
complains about a use of uninitialized data. Test case:
run valgrind ./txr -i and execute a (throw 'foo) expr.

* signal.c (sig_mask): If we are compiling with Valgrind
support, mark the old signal set defined just before passing
it to sigprocmask, so it has no uninitialized bits.