Change how error locations are reported in REPL.
Now that we have a multi-line mode, it no longer works that we
use the command line number as the parse line number, so that
an error from something that was parsed in command 42 comes
out as "line:42". This commit changes to the format
"expr-42:3": line 3 of expression 42. The expression number is
fixed in the parse name, and the line number is derived from
the line divisions in the entry itself.

* parser.c (repl): Dynamically compute the parse name,
such that it contains the command number. Then when parsing,
just leave the parse to begin at line 1.