C++ upkeep: cannot assign to volatile sigset_t.
This is a hack for an ugly problem. A plain old
struct is actually a class in C++, and assignment
to a struct goes thorugh a generated assignment
operator. Unfortunately, C++ doesn't generate
an assignment operator for volatile destinations,
so assignment to a volatile-qualified struct object
is erroneous.

* signal.h (copy_sigset): New inline function
to copy a sigset_t, even if the left hand side
is volatile. This works by stripping the
qualifier.
(sig_save_enable): Use the function.