Fix misleading documentation under if/iffi.
* txr.1: Examples under iff and iffi had two issues:
the false predicate function was used in a role which calls
for a function that unconditionally returns nil.
The correct function is nilf. Square brackets must
be used, otherwise nilf and identity are unbound
variables. Also the equivalence for [iff a] now
indicates that it's equiavlent to a, showing that
it's not useful to default both arguments.