Fix size growth of nested continuation capture.
The issue is that captures from within a restarted
continuation of the same continuation, are using the
restart prompt as the delimiter rather than re-using
captured block of the same name. Thus each successive
capture adds a new frame, causing the size to grow
drastically.

* unwind.c (uw_capture_cont): Allow the existing captured
block to be the delimiter; don't delimit up to
the revive_cont's prompt block of the same name.