Bugfix: structs not inheriting late static slot values.
This bug doesn't affect static slots that are defined in
defstruct, because those get initialized in the new type by
the static init function.

The bug is that the values of static slots added later with
static_slot_ensure are not inherited by subtypes that are
created later still.

(Since static_slot_ensure propagates a slot to subtypes which
already exist, those types get the slot value.)

* struct.c (make_struct_type): Copy the contents of the
static slot array of the supertype into the new type,
so that static slot values are inherited. We can just use
memcpy because the ordering of static slots is the same
in the new type, and the inherited ones precede the
new ones, due to the way the all_slots list is combined.

* txr.1: Clarify inheritance of static slots.