HTML doc: expand TOC section on jumpback.
When a section number is clicked, if the table of contents
section which contains it is collapsed, open it, and its
parents.

* genman.txr: (tocjump): New JS function.  add onclick action
to jump-back links, which calls tocjump JS function which does
the expansion.  The name anchor entries in the TOC are tagged
with a "tocanchor: class so they can be more specifically
selected by the tocjump function.