HTML doc: fix back jumping flaw.
Issue: wrong [+] changes to [-], leading to
inconsistent state.

* genman.txr: In TOC extraction and transformation,
we fix the stupid broken <DT>....<DD> that man2html puts
out and turn it into <DT>...</DT>.  This way a DL
now has a DT element as its previous sibling.
(tocjump): When we see a DL now, we get the correct
element which has the expand/collapse: the previous
sibling which is a DT containing the anchor.