debugger: stopped at line nil of nil.
Address some situations in which the debug_check call is given
a piece of spec with no source loc information attached.

* (h_var): Use rlcp when constructing new piece of spec
representing a substituted variable, or a spec confined
to just search for a piece of text with no trailing match.
(do_match_line): Pass the whole specline to debug_check,
rather than just the elem, which might just be a string
with no location info attached.
(match_files): Likewise.