Revert chr-isdigit/isxdigit, provide new functions.
It was a mistake to change the semantics of the return value
of chr-isdigit and chr-isdigit. It breaks code like
[partition-by chr-isdigit ...].  The behavior of chr-isdigit
and chr-isxdigit is restored to returning t and nil. New
chr-digit and chr-xdigit functions are introduced for
returning the digit value or nil.

* eval.c (eval_init): Register chr-digit and chr-xdigit
intrinsics.

* lib.c (chr_isdigit, chr_isxdigit): Restore old behavior.
(chr_digit, chr_xdigit): New functions.

* lib.h (chr_digit, chr_xdigit): Declared.

* txr.1: Everything documented.