random: wrong mask width for power-of-two moduli.
This mistake causes wasteful behavior for power-of-two
moduli in the random function, in both the bignum and
fixnum cases. For instance, the modulus 16 is taken
to be 17 bits wide. But we really want the width 16:
the number of bits needed for values in the range [0, 16).
The result isn't wrong, but the loop generates 17-bit
random numbers, and then throws away those which equal
or exceed the modulus, which is wasteful.

* mpi/mpi.c (mp_is_pow_two): New function.

* mpi/mpi.h (mp_is_pow_two): Declared.

* rand.c (random): In bignum case, after counting
bits in the modulus, subtract 1 if the modulus is a power
of two. In the fixnum case, subtract 1 from the modulus
and then count the bits in the reduced value.

* tests/013/maze.expected: regenerate due to different
prng behavior.