Workaround for apparent putc bug in Cygwin.
This fix is required for the stream socket test case to pass.
Some interaction between a stdio stream in line buffering
mode and the putc function causes a stream to lose data.
If we use fputs instead to output a character, the
issue goes away.

* stream.c (se_putc): When compiling for Cygwin, construct
a one-character-long string and use fputs, rather than putc.