Bugfix: support abstract UNIX socket addresses on Linux.
Making it work as already documented.

* socket.c (MIN): New macro.
(sockaddr_pack): Use utf8_dup_to_buf to convert
Unix socket path to a buffer of UTF-8 bytes, possibly with one
or more embedded null bytes. Copy as much of this as fits into
the sun_path member of struct sockaddr_un.

* txr.1: Improve documentation about the abstract names
on Linux.