Bugfix: track origin for built-in macros.
This fix is no longer that important, since due to the
previous commit, expand now tracks expansion origins
thoroughly, which hides this bug. However, expand_macro is
directly called from a few other functions, like macroexpand,
which don't benefit from that fix.

* eval.c (expand_macro): Add forgotten set_origin call
in the case when the native C expander is called.