Calculate sysroot even if executable is renamed.
If txr is renamed to something else in the same installation
directory, we can still calculate the sysroot.

* txr.c (substitute_basename): New function.
(sysroot_init): Make one more attempt to calculate the
sysroot by substituting the actual base name of
the program into the compile-time relative path, and
using that path agains the full path name.